+++ launchpy/config.py
@@ -5,11 +5,13 @@ config_object = {
     "api_key": "",
     "pathToKey": "",
     "secret": "",
-    "tokenEndpoint" : "https://ims-na1.adobelogin.com/ims/exchange/jwt",
+    "jwtTokenEndpoint" : "https://ims-na1.adobelogin.com/ims/exchange/jwt",
+    "oauthTokenEndpointV2" : "https://ims-na1.adobelogin.com/ims/token/v2",
     "date_limit" : 0,
     "scope_admin" : "https://ims-na1.adobelogin.com/s/ent_reactor_admin_sdk",
     "scope_dev" : "https://ims-na1.adobelogin.com/s/ent_reactor_sdk",
-    "official_scope" :  ""
+    "official_scope" :  "",
+    "scopes":""
 }
 
 header = {"Accept": "application/vnd.api+json;revision=1",